# swagger.yaml
...
openapi: 3.0.3
info:
  title: Clinic-AI API
  version: 1.0.0
  description: "AI-powered clinical intake system. This spec is being filled module-by-module."
servers:
  - url: http://localhost:8000
paths:
...
  /patients/:
    post:
      tags: [Patients]
      summary: Register a new patient and start intake session
      operationId: register_patient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterPatientRequest'
      responses:
        '201':
          description: Patient registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_RegisterPatientResponse'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Duplicate patient
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '422':
          description: Invalid symptom
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /patients/consultations/answer:
    post:
      tags: [Patients]
      summary: Answer an intake question
      operationId: answer_intake_question
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnswerIntakeRequest'
      responses:
        '200':
          description: Answer accepted, next question or completion returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_AnswerIntakeResponse'
        '400': {description: Validation error, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '404': {description: Patient or visit not found, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '409': {description: Intake already completed, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '422': {description: Question limit exceeded or duplicate question, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '500': {description: Internal server error, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
    patch:
      tags: [Patients]
      summary: Edit intake answer
      operationId: edit_intake_answer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditAnswerRequest'
      responses:
        '200': {description: Edit accepted, content: {application/json: {schema: {$ref: '#/components/schemas/ApiResponse_EditAnswerResponse'}}}}
        '400': {description: Validation error, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '404': {description: Patient or visit not found, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '500': {description: Internal server error, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
  /patients/webhook/images:
    post:
      tags: [Patients]
      summary: Upload medication images for intake
      operationId: upload_medication_images
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                patient_id: {type: string}
                visit_id: {type: string}
                images:
                  type: array
                  items:
                    type: string
                    format: binary
      responses:
        '200': {description: Successful uploads or errors }
        '400': {description: Validation error, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '500': {description: Internal server error, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
  /patients/{patient_id}/visits/{visit_id}/intake-images/{image_id}/content:
    get:
      tags: [Patients]
      summary: Get intake medication image content
      operationId: get_intake_medication_image_content
      parameters:
        - name: patient_id
          in: path
          required: true
          schema: {type: string}
        - name: visit_id
          in: path
          required: true
          schema: {type: string}
        - name: image_id
          in: path
          required: true
          schema: {type: string}
      responses:
        '200': {description: Raw content returned}
        '403': {description: Access denied, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '404': {description: Not found, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
  /patients/{patient_id}/visits/{visit_id}/images:
    get:
      tags: [Patients]
      summary: List medication images for a visit
      operationId: list_medication_images
      parameters:
        - name: patient_id
          in: path
          required: true
          schema: {type: string}
        - name: visit_id
          in: path
          required: true
          schema: {type: string}
      responses:
        '200': {description: JSON with image metadata}
        '404': {description: Not found, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '500': {description: Internal server error, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
  /patients/images/{image_id}:
    delete:
      tags: [Patients]
      summary: Delete a medication image
      operationId: delete_medication_image
      parameters:
        - name: image_id
          in: path
          required: true
          schema: {type: string}
      responses:
        '200': {description: Successful deletion}
        '404': {description: Not found, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '500': {description: Internal server error, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
  /patients/summary/previsit:
    post:
      tags: [Patients]
      summary: Generate pre-visit clinical summary
      operationId: generate_pre_visit_summary
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PreVisitSummaryRequest'
      responses:
        '200':
          description: Previsit summary generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_PreVisitSummaryResponse'
        '400': {description: Validation error, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '404': {description: Not found, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '422': {description: Intake not completed, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '500': {description: Internal server error, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
  /patients/{patient_id}/visits/{visit_id}/summary:
    get:
      tags: [Patients]
      summary: Get pre-visit summary
      operationId: get_pre_visit_summary
      parameters:
        - name: patient_id
          in: path
          required: true
          schema: {type: string}
        - name: visit_id
          in: path
          required: true
          schema: {type: string}
      responses:
        '200':
          description: Previsit summary found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_PreVisitSummaryResponse'
        '404': {description: Not found, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '500': {description: Internal server error, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
...
  /patients/{patient_id}/visits/{visit_id}/vitals:
    post:
      tags: [Patients]
      summary: Store vitals data for a visit
      operationId: store_vitals
      parameters:
        - name: patient_id
          in: path
          required: true
          schema:
            type: string
        - name: visit_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VitalsPayload'
      responses:
        '200':
          description: Vitals stored successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Dict'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Patient or visit not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    get:
      tags: [Patients]
      summary: Get vitals data for a visit
      operationId: get_vitals
      parameters:
        - name: patient_id
          in: path
          required: true
          schema:
            type: string
        - name: visit_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Get vitals success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_Dict'
        '404':
          description: Patient, visit, or vitals not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /notes/transcribe:
    post:
      tags: [notes]
      summary: Transcribe audio
      description: Queue audio transcription and return immediately (202).
      operationId: transcribe_audio
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [patient_id, visit_id, audio_file]
              properties:
                patient_id: {type: string}
                visit_id: {type: string}
                language: {type: string, default: en}
                audio_file: {type: string, format: binary}
      responses:
        '202': {description: Queued for transcription}
        '400': {description: Validation error, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '404': {description: Patient or visit not found, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '422': {description: Invalid audio file or visit status, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '500': {description: Internal server error, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
  /notes/soap/generate:
    post:
      tags: [notes]
      summary: Generate SOAP note
      description: Generate SOAP note from transcript.
      operationId: generate_soap_note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SoapGenerationRequest'
      responses:
        '200':
          description: SOAP note generated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SoapGenerationResponse'
        '400': {description: Validation error, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '404': {description: Patient or visit not found, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '422': {description: No transcript available or invalid visit status, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '500': {description: Internal server error, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
  /notes/{patient_id}/visits/{visit_id}/transcript:
    get:
      tags: [notes]
      summary: Get transcript
      description: Get transcript for a visit.
      operationId: get_transcript
      parameters:
        - name: patient_id
          in: path
          required: true
          schema: {type: string}
        - name: visit_id
          in: path
          required: true
          schema: {type: string}
      responses:
        '200':
          description: Transcript returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_TranscriptionSessionDTO'
        '404': {description: Transcript not found, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '422': {description: Invalid input, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
  /notes/{patient_id}/visits/{visit_id}/soap:
    get:
      tags: [notes]
      summary: Get SOAP note
      description: Get SOAP note for a visit.
      operationId: get_soap_note
      parameters:
        - name: patient_id
          in: path
          required: true
          schema: {type: string}
        - name: visit_id
          in: path
          required: true
          schema: {type: string}
      responses:
        '200':
          description: SOAP note returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_SoapNoteDTO'
        '404': {description: Patient, visit, or SOAP note not found, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '500': {description: Internal server error, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
  /notes/vitals:
    post:
      tags: [notes]
      summary: Store objective vitals
      operationId: store_vitals_notes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NotesVitalsPayload'
      responses:
        '200': {description: Vitals stored}
        '400': {description: Validation error, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '404': {description: Patient or visit not found, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '500': {description: Internal server error, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
  /notes/{patient_id}/visits/{visit_id}/vitals:
    get:
      tags: [notes]
      summary: Get vitals
      operationId: get_vitals_notes
      parameters:
        - name: patient_id
          in: path
          required: true
          schema: {type: string}
        - name: visit_id
          in: path
          required: true
          schema: {type: string}
      responses:
        '200':
          description: Vitals returned
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '404': {description: Patient or visit not found, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '500': {description: Internal server error, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
...
  /doctor/preferences:
    get:
      tags: [doctor]
      summary: Get doctor preferences
      description: Return latest preferences for the default doctor; independent of intake flow.
      operationId: get_doctor_preferences
      responses:
        '200':
          description: Doctor preferences loaded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_DoctorPreferencesResponse'
        '500': {description: Internal server error, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
    post:
      tags: [doctor]
      summary: Set doctor preferences
      description: Merge and persist doctor preferences; independent of intake flow.
      operationId: set_doctor_preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertPreferencesRequest'
      responses:
        '200':
          description: Preferences updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_DoctorPreferencesResponse'
        '500': {description: Internal server error, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
  /audio/:
    get:
      tags: [audio]
      summary: List audio files
      operationId: list_audio_files
      parameters:
        - name: patient_id
          in: query
          required: false
          schema: {type: string}
        - name: visit_id
          in: query
          required: false
          schema: {type: string}
        - name: audio_type
          in: query
          required: false
          schema: {type: string, description: 'adhoc | visit'}
        - name: limit
          in: query
          required: false
          schema: {type: integer, minimum: 1, maximum: 100, default: 50}
        - name: offset
          in: query
          required: false
          schema: {type: integer, minimum: 0, default: 0}
      responses:
        '200':
          description: Audio files listed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse_AudioListResponse'
        '500': {description: Failed to list audio files, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
...
  /audio/{audio_id}/download:
    get:
      tags: [audio]
      summary: Download audio file
      operationId: download_audio_file
      parameters:
        - name: audio_id
          in: path
          required: true
          schema: {type: string}
      responses:
        '200': {description: Binary stream}
        '404': {description: Audio not found, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
        '500': {description: Failed to download audio file, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
...
  /audio/stats/summary:
    get:
      tags: [audio]
      summary: Get audio storage statistics
      operationId: get_audio_stats
      responses:
        '200': {description: Stats returned, content: {application/json: {schema: {$ref: '#/components/schemas/ApiResponse_Dict'}}}}
        '500': {description: Failed to get audio statistics, content: {application/json: {schema: {$ref: '#/components/schemas/ErrorResponse'}}}}
...
components:
  schemas:
...
    ApiResponse_RegisterPatientResponse:
      type: object
      required: [success, message, timestamp, request_id, data]
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Created
        timestamp:
          type: string
          format: date-time
        request_id:
          type: string
          format: uuid
        data:
          $ref: '#/components/schemas/RegisterPatientResponse'
    RegisterPatientRequest:
      type: object
      required: [first_name, last_name, mobile, age, gender, recently_travelled, consent, country, language]
      properties:
        first_name:
          type: string
          minLength: 1
          maxLength: 40
        last_name:
          type: string
          minLength: 1
          maxLength: 40
        mobile:
          type: string
          pattern: "^(\\+|)[0-9]{8,16}$"
        age:
          type: integer
          minimum: 0
          maximum: 120
        gender:
          type: string
        recently_travelled:
          type: boolean
        consent:
          type: boolean
        country:
          type: string
          minLength: 2
          maxLength: 2
        language:
          type: string
          minLength: 2
          maxLength: 2
...